Quiet a lot of warnings

+++ src/daemon.c
@@ -47,7 +47,7 @@ int main (const int argc, char *argv[])
     _trace("main::writing_pid: %s", config.pid_path);
     config.wrote_pid = write_pid(config.pid_path);
 
-    const int count = arrlen(config.credentials);
+    const size_t count = arrlen(config.credentials);
     if (count == 0) { _warn("main::load_credentials: no credentials were loaded"); }
 
     struct state state = {0};
